cleanup

+++ oneshot.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
+set -eou pipefail
+
 cat << ENDL
       ____  ___   _     ____________
      / __ \/   | ( )  _/_/ ____/ __ \\ 
@@ -10,7 +12,14 @@ cat << ENDL
     50 Days In And Beyond
                                                                           
 ENDL
-insdir=$(mktemp -u)
-git clone --quiet https://github.com/day50-dev/sidechat $insdir
-$insdir/install.sh
-rm -r $insdir
+insdir=$(mktemp -d)
+if ! command -v lgit > /dev/null; then
+    curl -Ls https://github.com/day50-dev/tmux-ai-agent-helper/archive/refs/heads/main.zip -o $insdir/main.zip
+    cd $insdir && unzip main.zip
+    cd tmux-ai-agent-helper-main/ && ./install.sh
+else
+    rm -r $insdir
+    git clone --quiet https://github.com/day50-dev/sidechat $insdir
+    $insdir/install.sh
+fi
+rm -fr $insdir
